# api/swagger.yaml
swagger: '2.0'
info:
  description: API проекта "Голос поколения" для РКН и УМВД Липецкой области.
  version: 1.0.0
  title: API Voice Generation
  contact:
    email: antony.nor@yandex.ru
host: 'voice-gen-220-900.appspot.com'
basePath: /
...
schemes:
  - https
  - http
paths:
...
  /public/{imgName}:
      x-swagger-router-controller: Global
      get:
        tags:
          - global
        summary: Возвращает картинку из папки public
        operationId: publicGetImg
        produces:
          - image/png
        parameters:
          - in: path
            name: imgName
            required: true
            type: string
            description: Имя картинки
        responses:
          200:
            description: OK
            schema:
              type: file
              description: Возвращаемая картинка в png
          400:
            description: Invalid data
  /sendMsg:
    x-swagger-router-controller: Global
    post:
      tags:
        - global
      summary: Отправить сообщение
      operationId: sendMsg
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: ID диалога, ID отправителя и текст сообщения
          required: true
          schema:
            type: object
            required:
              - dlg_id
              - sender_id
              - msg_text
            properties:
              dlg_id:
                type: integer
                format: int32
                example: 4
              sender_id:
                type: integer
                format: int32
                example: 1
              msg_text:
                type: string
                example: >-
                  Здравстуйте! Подскажите, пожалуйста, лучший наркологический
                  диспансер в Липецке
            xml:
              name: parametres
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                example: OK
              msg_id:
                type: integer
                format: int32
                example: 5
            xml:
              name: response
        '400':
          description: Invalid data
  /getLastMsg:
    x-swagger-router-controller: Global
    post:
      tags:
        - global
      summary: Получить последнее сообщение в диалоге
      operationId: getLastMsg
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: ID диалога и ID отправителя
          required: true
          schema:
            type: object
            required:
              - dlg_id
              - sender_id
            properties:
              dlg_id:
                type: integer
                format: int32
                example: 4
              sender_id:
                type: integer
                format: int32
                example: 1
            xml:
              name: parameters
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                example: OK
              array_msg:
                type: array
                items:
                  $ref: '#/definitions/message'
            xml:
              name: response
        '400':
          description: Invalid data
  /getDialogMsgs:
    x-swagger-router-controller: Global
    post:
      tags:
        - global
      summary: Получить все сообщения диалога
      operationId: getDialogMsgs
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: ID диалога и ID отправителя
          required: true
          schema:
            type: object
            required:
              - dlg_id
              - sender_id
            properties:
              dlg_id:
                type: integer
                format: int32
                example: 4
              sender_id:
                type: integer
                format: int32
                example: 1
            xml:
              name: parameters
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                example: OK
              array_msg:
                type: array
                items:
                  $ref: '#/definitions/message'
            xml:
              name: response
        '400':
          description: Invalid data
  /getDialogInfo:
    x-swagger-router-controller: Global
    post:
      tags:
        - global
      summary: Получить информацию о диалоге
      operationId: getDialogInfo
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: ID диалога и ID отправителя
          required: true
          schema:
            type: object
            required:
              - dlg_id
              - sender_id
            properties:
              dlg_id:
                type: integer
                format: int32
                example: 4
              sender_id:
                type: integer
                format: int32
                example: 1
            xml:
              name: parameters
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                example: OK
              interlocutor:
                type: string
                example: Петров Александр
              role:
                type: string
                example: Volunteer
              last_msg:
                $ref: '#/definitions/message'
              create_dt:
                type: string
                format: date-time
                example: '2018-11-04T00:42:36-03:00'
            xml:
              name: response
        '400':
          description: Invalid data
  
  #
  # Admin/Volunteer
  #
...
definitions:
...
  message:
    type: object
    required:
      - msg_id
      - msg_text
      - msg_fromyou
      - msg_dt
    properties:
      msg_id:
        type: integer
        format: int32
        example: 5
      msg_text:
        type: string
        example: >-
          Здравстуйте! Подскажите, пожалуйста, лучший наркологический диспансер
          в Липецке
      msg_fromyou:
        type: boolean
        example: true
      msg_dt:
        type: string
        format: date-time
        example: '2018-11-04T00:42:36-03:00'
    xml:
      name: message
...